requirements.txt
streamlit>=1.35
python-dotenv>=1.0
supabase>=2.28.3,<3
psycopg2-binary>=2.9
